requirements.txt
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.0.0
notion-client==2.2.1
python-dotenv==1.0.1
